# api/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: Cinnabar API
  description: >-
    This is a Movies/Tv-series Reviews Server based on the OpenAPI 3.0
    specification. 
     If you want to know more about the project here are some useful links:
    - [The Tooren Romaros
    repository](https://github.com/CoPhos/ToorenRomaros/tree/master)

    - [The source API definition for the Tooren Romaros
    project](https://github.com/CoPhos/ToorenRomaros/blob/develop/tooren_romaros_api/src/main/resources/api/openapi.yaml)
  contact:
    name: Jorge Cobos
    url: https://www.JorgeCobos.com
    email: cobos1932@gmail.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: "1.0"
servers:
  - url: http://localhost:9090
    description: Development
  - url: http://localhost:9090
    description: Production
...
paths:
...
  /api/v1/streaming/film:
    post:
      tags:
        - post
        - Stream site
        - Film
      summary: Create Film-Streaming Site Relationship
      description: >-
        This endpoint allows users to create a new relationship between a film
        and a streaming site by providing the necessary details using the
        provided payload.
      operationId: createStreamSiteFilm
      parameters:
        - name: CreateStreamSiteFilmDto
          in: query
          description: CreateStreamSiteFilmDto object
          required: true
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateStreamSiteFilmDto"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStreamSiteFilmDto"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetStreamSiteFilmDto"
        "400":
          description: Bad Request
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
      security:
        - Bearer Authentication:
            - ADMIN
...
  /api/v1/films:
    get:
      tags:
        - Film
        - get
      summary: Get Film by ID
      description: >-
        This endpoint retrieves a specific film identified by its unique
        identifier (ID).
      operationId: getFilmByDynamicQuery
      parameters:
        - name: streamId
          in: query
          description: List of stream IDs to filter films by.
          required: false
          schema:
            type: array
            items:
              type: string
        - name: genres
          in: query
          description: List of genres to filter films by.
          required: false
          schema:
            type: array
            items:
              type: string
        - name: suitableFor
          in: query
          description: List of suitability factors to filter films by.
          required: false
          schema:
            type: array
            items:
              type: string
          example: "[PG,PG-18]"
        - name: filmType
          in: query
          description: >-
            Filter films by type (e.g., "all", "1", "2"). Default is "all". (1
            is for movies, 2 is for series).
          required: false
          schema:
            type: string
            default: all
        - name: atTheaters
          in: query
          description: Filter films currently at theaters or not. Default is "false".
          required: false
          schema:
            type: string
        - name: commingSoonStreaming
          in: query
          description: >-
            Filter films coming soon on streaming platforms or not. Default is
            "false".
          required: false
          schema:
            type: string
        - name: commingSoonTheaters
          in: query
          required: false
          schema:
            type: string
        - name: atStreaming
          in: query
          description: Filter films currently streaming or not. Default is "false".
          required: false
          schema:
            type: string
        - name: orderBy
          in: query
          description: Array of fields for sorting films. Default is ["tittle-asc"]
          required: false
          schema:
            type: array
            items:
              type: string
            default:
              - tittle-asc
          example: "[title-asc, releaseDate-desc]"
        - name: userRating
          in: query
          description: >-
            Filter films by user rating. "down" for ratings less than 60
            otherwise "up"
          required: false
          schema:
            type: string
        - name: superRating
          in: query
          description: >-
            Filter films by super rating. "down" for ratings less than 60
            otherwise "up"
          required: false
          schema:
            type: string
        - name: search
          in: query
          description: " Perform a search query to filter films by title."
          required: false
          schema:
            type: string
        - name: page
          in: query
          description: The number page of the request.
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          description: The page size of each request. Default is "6"
          required: false
          schema:
            type: integer
            format: int32
            default: 6
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetDynamicQyeryFilmDto"
        "400":
          description: Bad Request
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
    post:
      tags:
        - post
        - Film
      summary: Create Film
      description: >-
        This endpoint allows users to create a new film by providing the
        necessary details using the provided payload.
      operationId: createFilm
      parameters:
        - name: createFilmDto
          in: query
          description: createFilmDto object
          required: true
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateFilmDto"
      requestBody:
        content:
          application/json:
            schema:
              description: createFilmDto object
              oneOf:
                - $ref: "#/components/schemas/CreateFilmDto"
                - $ref: "#/components/schemas/CreateMovieDto"
                - $ref: "#/components/schemas/CreateSerieDto"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetFilmDto"
        "400":
          description: Bad Request
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
      security:
        - Bearer Authentication:
            - ADMIN
...
  /api/v1/films/{filmId}:
    get:
      tags:
        - Film
        - get
      summary: Get Film by ID
      description: >-
        This endpoint retrieves a specific film identified by its unique
        identifier (ID).
      operationId: getFilm
      parameters:
        - name: filmId
          in: path
          description: ID of the Film
          required: true
          schema:
            pattern: >-
              [a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetFilmDto"
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
    delete:
      tags:
        - Film
        - delete
      summary: Delete Film
      description: >-
        This endpoint allows users to delete a specific film identified by its
        unique identifier (filmId). Deleting a film is a permanent action and
        cannot be undone.
      operationId: deleteFilm
      parameters:
        - name: filmId
          in: path
          description: The ID of the Film
          required: true
          schema:
            pattern: >-
              [a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}
            type: string
      responses:
        "202":
          description: Accepted
          content:
            text/plain: {}
        "401":
          description: Unauthorized
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
      security:
        - Bearer Authentication:
            - ADMIN
    patch:
      tags:
        - patch
        - Film
      summary: Update Film
      description: >-
        This endpoint allows users to update a specific film identified by its
        unique identifier (ID). Users can modify the properties of the film
        using the provided update payload.
      operationId: updateFilm
      parameters:
        - name: filmId
          in: path
          description: ID of the Film
          required: true
          schema:
            pattern: >-
              [a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}
            type: string
        - name: updateFilmDto
          in: query
          description: updateFilmDto object
          required: true
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateFilmDto"
      requestBody:
        content:
          application/json:
            schema:
              description: updateFilmDto object
              oneOf:
                - $ref: "#/components/schemas/UpdateFilmDto"
                - $ref: "#/components/schemas/UpdateMovieDto"
                - $ref: "#/components/schemas/UpdateSerieDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetFilmDto"
        "400":
          description: Bad Request
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - Bearer Authentication:
            - ADMIN
...
  /api/v1/users/{userId}/comments/films/{filmId}:
    get:
      tags:
        - Comment
        - get
      summary: Get Comment by Film and User
      description: >-
        This endpoint retrieves a specific comment made by a user in a
        particular film. The comment is identified based on the unique
        identifiers of both the user (userId) and the film (filmId).
      operationId: getCommentByFilmIdAndUserId
      parameters:
        - name: userId
          in: path
          description: ID of the user
          required: true
          schema:
            pattern: >-
              [a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}
            type: string
        - name: filmId
          in: path
          description: ID ofthe film
          required: true
          schema:
            pattern: >-
              [a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}
            type: string
        - name: reported
          in: query
          description: >-
            Filter comments based on whether they are reported or not. Default
            is "false".
          required: false
          schema:
            type: string
            default: "false"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetCommentDto"
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
  /api/v1/sagas/{sagaId}/films:
    get:
      tags:
        - Film
        - get
      summary: Get Film by ID
      description: >-
        This endpoint retrieves all films associated with a specific saga
        identified by its unique identifier (ID).
      operationId: getAllFilmBySagaId
      parameters:
        - name: sagaId
          in: path
          description: ID of the Saga
          required: true
          schema:
            pattern: >-
              [a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetFilmDto"
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
...
  /api/v1/films/{filmId}/streamingsites:
    get:
      tags:
        - Stream site
        - Film
        - get
      summary: Get All Streaming sites From a Specific Film
      description: >-
        This endpoint retrieves all entities of the Film-Streaming relationship
        for a specific film identified by its unique identifier (filmId).
      operationId: getStreamSiteFilm
      parameters:
        - name: filmId
          in: path
          description: ID of the Film
          required: true
          schema:
            pattern: >-
              [a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetStreamSiteFilmDto"
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
...
  /api/v1/streaming/film/{relationshipId}:
    delete:
      tags:
        - Stream site
        - Film
        - delete
      summary: Delete Film-Streaming Relationship
      description: >-
        This endpoint allows users to delete a specific film-Streaming
        relationship identified by its unique identifier (relationshipId).
        Deleting a relationship is a permanent action and cannot be undone.
      operationId: deleteStreamSiteFilm
      parameters:
        - name: relationshipId
          in: path
          description: The ID of the Relationship
          required: true
          schema:
            pattern: >-
              [a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}
            type: string
      responses:
        "202":
          description: Accepted
          content:
            text/plain: {}
        "401":
          description: Unauthorized
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
      security:
        - Bearer Authentication:
            - ADMIN
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        errorCode:
          type: string
        message:
          type: string
        status:
          type: integer
          format: int32
        url:
          type: string
        reqMethod:
          type: string
        timestamp:
          type: string
          format: date-time
...
    CreateStreamSiteFilmDto:
      required:
        - buy
        - rent
        - url
      type: object
      properties:
        url:
          maxLength: 255
          minLength: 0
          type: string
        rent:
          type: boolean
        buy:
          type: boolean
        filmId:
          type: string
        streamsiteId:
          type: string
    GetStreamSiteFilmDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
        rent:
          type: boolean
        buy:
          type: boolean
        filmName:
          type: string
        streamSiteName:
          type: string
        streamsiteId:
          type: string
          format: uuid
...
    CreateFilmDto:
      required:
        - filmType
        - tittle
      type: object
      properties:
        tittle:
          maxLength: 255
          minLength: 0
          type: string
        synopsis:
          maxLength: 255
          minLength: 0
          type: string
        originalLanguage:
          maxLength: 3
          minLength: 0
          type: string
        distributor:
          maxLength: 50
          minLength: 0
          type: string
        suitableFor:
          maxLength: 5
          minLength: 0
          type: string
        coomingSoon:
          type: string
          format: date
        streamingReleaseDate:
          type: string
          format: date
        filmType:
          type: integer
          format: int32
        sagaId:
          type: string
        averageSuperRating:
          type: number
          format: float
        averageUserRating:
          type: number
          format: float
        criticsConsensus:
          type: string
        aspectRatio:
          type: string
        soundMix:
          type: string
        viewCount:
          type: integer
          format: int64
        addedDate:
          type: string
          format: date
      discriminator:
        propertyName: filmType
    CreateMovieDto:
      required:
        - filmType
        - tittle
      type: object
      allOf:
        - $ref: "#/components/schemas/CreateFilmDto"
        - type: object
          properties:
            boxOffice:
              maxLength: 15
              minLength: 0
              type: string
            runTime:
              maxLength: 5
              minLength: 0
              type: string
            theatersReleaseDate:
              type: string
              format: date
            atTheaters:
              type: boolean
    CreateSerieDto:
      required:
        - filmType
        - tittle
      type: object
      allOf:
        - $ref: "#/components/schemas/CreateFilmDto"
        - type: object
          properties:
            seasonNumber:
              type: string
    GetFilmDto:
      type: object
      properties:
        id:
          type: string
        tittle:
          type: string
        synopsis:
          type: string
        originalLanguage:
          type: string
        distributor:
          type: string
        suitableFor:
          type: string
        coomingSoon:
          type: string
          format: date
        streamingReleaseDate:
          type: string
          format: date
        filmType:
          type: integer
          format: int32
        sagaId:
          type: string
        averageSuperRating:
          type: number
          format: float
        averageUserRating:
          type: number
          format: float
        criticsConsensus:
          type: string
        aspectRatio:
          type: string
        soundMix:
          type: string
        viewCount:
          type: integer
          format: int64
        addedDate:
          type: string
          format: date
        mainImageId:
          type: string
      discriminator:
        propertyName: filmType
...
    GetCommentDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        body:
          type: string
        publicationDateTime:
          type: string
          format: date-time
        likeCount:
          type: integer
          format: int32
        reported:
          type: boolean
        userId:
          type: string
        username:
          type: string
        filmId:
          type: string
        rating:
          type: integer
          format: int32
        spoiler:
          type: boolean
        filmName:
          type: string
        mainImageId:
          type: string
        theatersReleaseDate:
          type: string
...
    UpdateFilmDto:
      required:
        - filmType
      type: object
      properties:
        tittle:
          maxLength: 255
          minLength: 0
          type: string
        synopsis:
          maxLength: 255
          minLength: 0
          type: string
        originalLanguage:
          maxLength: 3
          minLength: 0
          type: string
        distributor:
          maxLength: 50
          minLength: 0
          type: string
        suitableFor:
          maxLength: 5
          minLength: 0
          type: string
        coomingSoon:
          type: string
          format: date
        streamingReleaseDate:
          type: string
          format: date
        filmType:
          type: integer
          format: int32
        sagaId:
          type: string
        averageSuperRating:
          type: number
          format: float
        averageUserRating:
          type: number
          format: float
        criticsConsensus:
          type: string
        aspectRatio:
          type: string
        soundMix:
          type: string
        viewCount:
          type: integer
          format: int64
        addedDate:
          type: string
          format: date
      discriminator:
        propertyName: filmType
    UpdateMovieDto:
      required:
        - filmType
      type: object
      allOf:
        - $ref: "#/components/schemas/UpdateFilmDto"
        - type: object
          properties:
            boxOffice:
              maxLength: 15
              minLength: 0
              type: string
            runTime:
              maxLength: 5
              minLength: 0
              type: string
            theatersReleaseDate:
              type: string
              format: date
            atTheaters:
              type: boolean
    UpdateSerieDto:
      required:
        - filmType
      type: object
      allOf:
        - $ref: "#/components/schemas/UpdateFilmDto"
        - type: object
          properties:
            seasonNumber:
              type: string
...
    GetDynamicQyeryFilmDto:
      type: object
      properties:
        id:
          type: string
        tittle:
          type: string
        coomingSoon:
          type: string
          format: date
        streamingReleaseDate:
          type: string
          format: date
        filmType:
          type: integer
          format: int32
        averageSuperRating:
          type: number
          format: float
        averageUserRating:
          type: number
          format: float
        viewCount:
          type: integer
          format: int64
        mainImageId:
          type: string
      discriminator:
        propertyName: filmType
  securitySchemes:
    Bearer Authentication:
      type: http
      scheme: bearer
      bearerFormat: JWT